requirements.txt
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
databases==0.5.5
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.10
django-mongoengine==0.5.4
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.2
filelock==3.6.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
jwt==1.3.1
mongoengine==0.24.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
passlib==1.7.4
Pillow==9.1.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.7
pymongo==3.12.1
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
python-snappy==0.6.1
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.35
sqlparse==0.2.4
starlette==0.17.1
toml==0.10.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.8
users==1.0.dev0
uvicorn==0.17.6
virtualenv==20.14.1
watchgod==0.8.2
websockets==10.3
winkerberos==0.9.0
wrapt==1.14.0
